chore: add prettier and contribution/code of conduct files (#4)
chore: add prettier and contribution/code of conduct files


This adds:

    A code of conduct
    A contribution guide
    prettier to enforce formatting
    Pull request template

Why formatting?

I don't want to mess around with style guides and linters. Using a formatter allows us to have code that is consistent in some key ways and requires little overhead.

The formatter code won't be 100% to everyone's taste, but at least it will be mostly inoffensive and above all consistent.